refactor(node-exec): merge node exec types into owners
