+++ crates/gateway/Cargo.toml
@@ -37,7 +37,6 @@ moltis-memory          = { workspace = true }
 moltis-metrics         = { optional = true, workspace = true }
 moltis-msteams         = { workspace = true }
 moltis-network-filter  = { features = ["proxy", "service"], optional = true, workspace = true }
-moltis-node-exec-types = { workspace = true }
 moltis-nostr           = { optional = true, workspace = true }
 moltis-oauth           = { workspace = true }
 moltis-onboarding      = { workspace = true }
