+++ crates/gateway/src/node_exec.rs
@@ -28,12 +28,90 @@ use crate::{
     state::GatewayState,
 };
 
-// Re-export core node execution types from the dedicated crate.
-pub use moltis_node_exec_types::{
-    BLOCKED_ENV_PREFIXES, NodeExecResult, NodeInfo, NodeInfoProvider, NodeProviderInfo,
-    SAFE_ENV_ALLOWLIST, SAFE_ENV_PREFIX_ALLOWLIST, SSH_ID_PREFIX, SSH_TARGET_ID_PREFIX, filter_env,
-    is_safe_env, is_valid_env_key, ssh_node_id, ssh_stored_node_id, ssh_target_matches,
-};
+use moltis_tools::nodes::{NodeInfo, NodeInfoProvider, NodeProviderInfo};
+
+/// Result of a remote command execution on a node.
+#[derive(Debug, Clone, Default, serde::Serialize, serde::Deserialize)]
+pub struct NodeExecResult {
+    pub stdout: String,
+    pub stderr: String,
+    pub exit_code: i32,
+}
+
+/// Environment variables that are safe to forward to a remote node.
+pub const SAFE_ENV_ALLOWLIST: &[&str] = &["TERM", "LANG", "COLORTERM", "NO_COLOR", "FORCE_COLOR"];
+
+/// Environment variable prefixes that are safe to forward.
+pub const SAFE_ENV_PREFIX_ALLOWLIST: &[&str] = &["LC_"];
+
+/// Environment variable patterns that must NEVER be forwarded to a remote node.
+pub const BLOCKED_ENV_PREFIXES: &[&str] = &[
+    "DYLD_",
+    "LD_",
+    "NODE_OPTIONS",
+    "PYTHON",
+    "PERL",
+    "RUBYOPT",
+    "SHELLOPTS",
+    "PS4",
+    "MOLTIS_",
+    "OPENAI_",
+    "ANTHROPIC_",
+    "AWS_",
+    "GOOGLE_",
+    "AZURE_",
+];
+
+/// SSH node ID prefix.
+pub const SSH_ID_PREFIX: &str = "ssh:";
+
+/// SSH target ID prefix.
+pub const SSH_TARGET_ID_PREFIX: &str = "ssh:target:";
+
+pub fn ssh_node_id(target: &str) -> String {
+    format!("{SSH_ID_PREFIX}{target}")
+}
+
+pub fn ssh_stored_node_id(id: i64) -> String {
+    format!("{SSH_TARGET_ID_PREFIX}{id}")
+}
+
+pub fn ssh_target_matches(node_ref: &str, target: &str) -> bool {
+    node_ref == "ssh" || node_ref == target || node_ref.strip_prefix(SSH_ID_PREFIX) == Some(target)
+}
+
+pub fn filter_env(env: &HashMap<String, String>) -> HashMap<String, String> {
+    env.iter()
+        .filter(|(key, _)| is_safe_env(key) && is_valid_env_key(key))
+        .map(|(key, value)| (key.clone(), value.clone()))
+        .collect()
+}
+
+pub fn is_safe_env(key: &str) -> bool {
+    for prefix in BLOCKED_ENV_PREFIXES {
+        if key.starts_with(prefix) {
+            return false;
+        }
+    }
+
+    if SAFE_ENV_ALLOWLIST.contains(&key) {
+        return true;
+    }
+
+    for prefix in SAFE_ENV_PREFIX_ALLOWLIST {
+        if key.starts_with(prefix) {
+            return true;
+        }
+    }
+
+    false
+}
+
+pub fn is_valid_env_key(key: &str) -> bool {
+    let mut chars = key.chars();
+    matches!(chars.next(), Some(ch) if ch.is_ascii_alphabetic() || ch == '_')
+        && chars.all(|ch| ch.is_ascii_alphanumeric() || ch == '_')
+}
 
 pub(crate) fn ssh_node_info(target: &str) -> NodeInfo {
     NodeInfo {
