+++ crates/gateway/src/server/prepare_core/post_state.rs
@@ -803,7 +803,7 @@ pub(super) async fn complete_startup(
         }
 
         {
-            let node_info_provider: Arc<dyn moltis_node_exec_types::NodeInfoProvider> =
+            let node_info_provider: Arc<dyn moltis_tools::nodes::NodeInfoProvider> =
                 Arc::new(crate::node_exec::GatewayNodeInfoProvider::new(
                     Arc::clone(&state),
                     config.tools.exec.ssh_target.clone(),
