--- crates/node-exec-types/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name                 = "moltis-node-exec-types"
-version.workspace    = true
-edition.workspace    = true
-repository.workspace = true
-
-[dependencies]
-anyhow         = { workspace = true }
-async-trait    = { workspace = true }
-moltis-metrics = { optional = true, workspace = true }
-serde          = { workspace = true, features = ["derive"] }
-tracing        = { optional = true, workspace = true }
-
-[features]
-default = ["metrics", "tracing"]
-metrics = ["dep:moltis-metrics"]
-tracing = ["dep:tracing"]
-
-[lints]
-workspace = true
